refactor: fix merge conflict with different location of location-selector
