+++ angular-app/src/app/app.config.ts
@@ -1,7 +1,7 @@
 import { ApplicationConfig, provideBrowserGlobalErrorListeners, provideZoneChangeDetection } from '@angular/core';
 import { provideRouter } from '@angular/router';
-import { LocationSelectorService } from './location-selector/location-selector-service';
-import { NominatimLocationSelectorService } from './location-selector/nominatim-location-selector.service';
+import { LocationSelectorService } from './components/location-selector/location-selector-service';
+import { NominatimLocationSelectorService } from './components/location-selector/nominatim-location-selector.service';
 
 import { routes } from './app.routes';
