--- angular-app/src/app/components/location-choice-component/location-choice-component.html
@@ -1,26 +1,0 @@
-<div class="location-choice">
-  <h3>Choose Location</h3>
-  <div class="form-group">
-    <label for="country">Country</label>
-    <input type="text" id="country" name="country" placeholder="Search country...">
-    <div class="dropdown">
-      <!-- Country dropdown options -->
-    </div>
-  </div>
-  <div class="form-group">
-    <label for="state">State</label>
-    <input type="text" id="state" name="state" placeholder="Search state...">
-    <div class="dropdown">
-      <!-- State dropdown options -->
-    </div>
-  </div>
-  <div class="form-group">
-    <label for="locality">Locality</label>
-    <input type="text" id="locality" name="locality" placeholder="Search locality...">
-    <div class="dropdown">
-      <!-- Locality dropdown options -->
-    </div>
-  </div>
-  <button type="button">Search</button>
-</div>
-
