--- angular-app/src/app/components/location-choice-component/location-choice-component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-location-choice-component',
-  imports: [],
-  templateUrl: './location-choice-component.html',
-  styleUrl: './location-choice-component.scss'
-})
-export class LocationChoiceComponent {
-
-}
