+++ angular-app/src/app/components/location-selector/location-option.ts
