+++ angular-app/src/app/components/location-selector/location-selector-service.ts
