+++ angular-app/src/app/components/location-selector/nominatim-location-selector.service.ts
