+++ angular-app/src/app/components/post-edit/post-edit.html
@@ -17,7 +17,7 @@
     </div>
 
     <div class="location-section">
-      <app-location-choice-component></app-location-choice-component>
+      <app-location-selector></app-location-selector>
     </div>
 
     <div class="form-actions">
