+++ angular-app/src/app/components/post-edit/post-edit.ts
@@ -1,10 +1,10 @@
 import { Component } from '@angular/core';
 import { ImageUpload } from '../image-upload/image-upload';
-import { LocationChoiceComponent } from '../location-choice-component/location-choice-component';
+import { LocationSelector } from '../location-selector/location-selector.component';
 
 @Component({
   selector: 'app-post-edit',
-  imports: [ImageUpload, LocationChoiceComponent],
+  imports: [ImageUpload, LocationSelector],
   templateUrl: './post-edit.html',
   styleUrl: './post-edit.scss'
 })
