+++ angular-app/src/app/components/post-filter-search-panel/post-filter-search-panel.html
@@ -1,7 +1,7 @@
 <div class="post-filter-search-panel">
   <h2>Filter Posts</h2>
 
-  <app-location-choice-component></app-location-choice-component>
+  <app-location-selector></app-location-selector>
 
   <div class="filter-options">
     <div class="form-group">
