+++ angular-app/src/app/components/post-filter-search-panel/post-filter-search-panel.ts
@@ -1,9 +1,9 @@
 import { Component } from '@angular/core';
-import { LocationChoiceComponent } from '../location-choice-component/location-choice-component';
+import { LocationSelector } from '../location-selector/location-selector.component';
 
 @Component({
   selector: 'app-post-filter-search-panel',
-  imports: [LocationChoiceComponent],
+  imports: [LocationSelector],
   templateUrl: './post-filter-search-panel.html',
   styleUrl: './post-filter-search-panel.scss'
 })
