+++ angular-app/src/app/components/profile-editor/profile-editor.html
@@ -14,7 +14,7 @@
       <textarea id="description" name="description" rows="4"></textarea>
     </div>
     <div class="form-group">
-      <app-location-choice-component></app-location-choice-component>
+      <app-location-selector></app-location-selector>
     </div>
     <button type="submit">Save Changes</button>
   </form>
