+++ angular-app/src/app/components/profile-editor/profile-editor.ts
@@ -1,9 +1,9 @@
 import { Component } from '@angular/core';
-import { LocationChoiceComponent } from '../location-choice-component/location-choice-component';
+import { LocationSelector } from '../location-selector/location-selector.component';
 
 @Component({
   selector: 'app-profile-editor',
-  imports: [LocationChoiceComponent],
+  imports: [LocationSelector],
   templateUrl: './profile-editor.html',
   styleUrl: './profile-editor.scss'
 })
