+++ angular-app/src/app/components/sign-up-component/sign-up-component.html
@@ -1,6 +1,6 @@
 <div class="sign-up">
   <h2>Sign Up</h2>
-  <app-location-choice-component></app-location-choice-component>
+  <app-location-selector></app-location-selector>
   <form>
     <div class="form-group">
       <label for="firstName">First Name</label>
