+++ angular-app/src/app/components/sign-up-component/sign-up-component.ts
@@ -1,9 +1,9 @@
 import { Component } from '@angular/core';
-import { LocationChoiceComponent } from '../location-choice-component/location-choice-component';
+import { LocationSelector } from '../location-selector/location-selector.component';
 
 @Component({
   selector: 'app-sign-up-component',
-  imports: [LocationChoiceComponent],
+  imports: [LocationSelector],
   templateUrl: './sign-up-component.html',
   styleUrl: './sign-up-component.scss'
 })
