+++ angular-app/src/app/pages/create-post-page/create-post-page.html
@@ -3,7 +3,7 @@
 
   <form>
     <div class="location-section">
-      <app-location-choice-component></app-location-choice-component>
+      <app-location-selector></app-location-selector>
 
       <div class="map-section">
         <h3>Choose Location on Map</h3>
