+++ angular-app/src/app/pages/create-post-page/create-post-page.ts
@@ -1,11 +1,11 @@
 import { Component } from '@angular/core';
-import { LocationChoiceComponent } from '../../components/location-choice-component/location-choice-component';
+import { LocationSelector } from '../../components/location-selector/location-selector.component';
 import { Map } from '../../components/map/map';
 import { ImageUpload } from '../../components/image-upload/image-upload';
 
 @Component({
   selector: 'app-create-post-page',
-  imports: [LocationChoiceComponent, Map, ImageUpload],
+  imports: [LocationSelector, Map, ImageUpload],
   templateUrl: './create-post-page.html',
   styleUrl: './create-post-page.scss'
 })
